Create WorkWithMongoInTheCloud.js
Instructions and working demo for the MongoDB lab. The setup instructions in this file are brief comments, to be used in conjunction with the provided PDF instructions. We need to register with MongoDB Atlas, not mLab as in the PDF, because mLab was acquired by MongoDB, Inc., so the cloud registration and initial setup are different.